+++ NEWS.md
@@ -1,6 +1,7 @@
 # latexdiffr (development version)
 
 * Added a `compile` option to (not) compile the diff.tex file to pdf
+* Added support for quarto files
 
 # latexdiffr 0.1.0